parent_path cleaned/eliminated, application controller methods in clearer modules
