+++ app/views/articles/_form.html.erb
@@ -1,18 +1,7 @@
 <%= simple_form_for(@article) do |f| %>
-  <% if @article.errors.any? %>
-    <div id="error_explanation">
-      <h2><%= pluralize(@article.errors.count, "error") %> prohibited this article from being saved:</h2>
-
-      <ul>
-      <% @article.errors.full_messages.each do |msg| %>
-        <li><%= msg %></li>
-      <% end %>
-      </ul>
-    </div>
-  <% end %>
-    <%= f.input :title %><br />
+	<%= f.input :title %><br />
     <%= f.input :link %><br />
-  <div class="actions">
-    <%= f.submit %>
-  </div>
+  	<div class="actions">
+    	<%= f.submit %>
+  	</div>
 <% end %>
