+++ app/views/comments/show.html.erb
@@ -1,7 +1,6 @@
 <p>
 	<strong><%= link_to @comment.username, user_path(@comment.user) %></strong>
-	<em>on <%= @comment.created_at.to_s(:mdyt) %> | <%= link_to "parent", @comment.commentable %></em><br />
-	<%= @comment.content %><br />
+	<em>on <%= @comment.created_at.to_s(:mdyt) %> | <%= link_to "parent", @comment.commentable %><br />
 </p>
 
 <%= render "comments/form"%>
